# openapi.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - App
  /customers/me/onboarding:
    post:
      operationId: CustomersController_onboarding
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerOnboardingDto'
      responses:
        '201':
          description: ''
      tags: &ref_0
        - Customers
  /customers/me/profile:
    patch:
      operationId: CustomersController_updateProfile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerProfileDto'
      responses:
        '200':
          description: ''
      tags: *ref_0
  /customers/me/merchants:
    post:
      operationId: CustomersController_associateMerchant
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssociateMerchantDto'
      responses:
        '201':
          description: ''
      tags: *ref_0
    get:
      operationId: CustomersController_listCustomerMerchants
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_0
  /customers/me/merchants/{merchantId}:
    patch:
      operationId: CustomersController_unsubscribeMerchant
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnsubscribeMerchantDto'
      responses:
        '200':
          description: ''
      tags: *ref_0
  /merchants/config/{slug}:
    get:
      operationId: MerchantsController_getConfigBySlug
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: &ref_1
        - Merchants
  /merchants/me:
    get:
      operationId: MerchantsController_listMyMerchants
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags: *ref_1
  /merchants/onboarding:
    post:
      operationId: MerchantsController_onboarding
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMerchantOnboardingDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      tags: *ref_1
  /merchants/{merchantId}/onboarding:
    patch:
      operationId: MerchantsController_updateMerchantOnboarding
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMerchantOnboardingDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags: *ref_1
  /merchants/{merchantId}/staff:
    post:
      operationId: MerchantsController_createMerchantStaff
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMerchantStaffDto'
      responses:
        '201':
          description: ''
      tags: *ref_1
    get:
      operationId: MerchantsController_listMerchantStaff
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
  /merchants/{merchantId}/customers:
    get:
      operationId: MerchantsController_listMerchantCustomers
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
  /merchants/{merchantId}/customers/{customerId}/status:
    patch:
      operationId: MerchantsController_updateMerchantCustomerStatus
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
        - name: customerId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMerchantCustomerStatusDto'
      responses:
        '200':
          description: ''
      tags: *ref_1
  /merchants/{merchantId}/staff/{staffId}:
    get:
      operationId: MerchantsController_getStaffDetail
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
        - name: staffId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
    patch:
      operationId: MerchantsController_updateMerchantStaff
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
        - name: staffId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMerchantStaffDto'
      responses:
        '200':
          description: ''
      tags: *ref_1
    delete:
      operationId: MerchantsController_removeMerchantStaff
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
        - name: staffId
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      tags: *ref_1
  /merchants/{merchantId}/staff/{staffId}/services:
    put:
      operationId: MerchantsController_setStaffServices
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
        - name: staffId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetStaffServicesDto'
      responses:
        '200':
          description: ''
      tags: *ref_1
  /merchants/{merchantId}/staff/{staffId}/working-hours:
    put:
      operationId: MerchantsController_setStaffWorkingHours
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
        - name: staffId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetStaffWorkingHoursDto'
      responses:
        '200':
          description: ''
      tags: *ref_1
  /merchants/{merchantId}/logo:
    put:
      operationId: MerchantsController_uploadLogo
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags: *ref_1
    delete:
      operationId: MerchantsController_deleteLogo
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      security:
        - bearer: []
      tags: *ref_1
  /merchants/{merchantId}/services:
    post:
      operationId: ServicesController_create
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceDto'
      responses:
        '201':
          description: ''
      tags: &ref_2
        - Services
    get:
      operationId: ServicesController_findAll
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
  /merchants/{merchantId}/services/{serviceId}:
    get:
      operationId: ServicesController_findOne
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
        - name: serviceId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
    patch:
      operationId: ServicesController_update
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
        - name: serviceId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceDto'
      responses:
        '200':
          description: ''
      tags: *ref_2
    delete:
      operationId: ServicesController_remove
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
        - name: serviceId
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      tags: *ref_2
  /appointments/availability:
    get:
      operationId: AppointmentsController_getAvailability
      parameters: []
      responses:
        '200':
          description: ''
      tags: &ref_3
        - Appointments
  /appointments/available-slots:
    get:
      operationId: AppointmentsController_getAvailableSlots
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_3
  /appointments/day-calendar:
    get:
      operationId: AppointmentsController_getDayCalendar
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_3
  /appointments/staff-day-calendar:
    get:
      operationId: AppointmentsController_getStaffDayCalendar
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_3
  /merchants/{merchantId}/appointments:
    post:
      operationId: AppointmentsController_create
      parameters:
        - name: merchantId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentDto'
      responses:
        '201':
          description: ''
      tags: *ref_3
info:
  title: Business Engine API
  description: API documentation
  version: 1.0.0
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CustomerOnboardingDto:
      type: object
      properties: {}
    UpdateCustomerProfileDto:
      type: object
      properties: {}
    AssociateMerchantDto:
      type: object
      properties: {}
    UnsubscribeMerchantDto:
      type: object
      properties: {}
    CreateMerchantOnboardingDto:
      type: object
      properties: {}
    UpdateMerchantOnboardingDto:
      type: object
      properties: {}
    CreateMerchantStaffDto:
      type: object
      properties: {}
    UpdateMerchantCustomerStatusDto:
      type: object
      properties: {}
    UpdateMerchantStaffDto:
      type: object
      properties: {}
    SetStaffServicesDto:
      type: object
      properties: {}
    SetStaffWorkingHoursDto:
      type: object
      properties: {}
    CreateServiceDto:
      type: object
      properties: {}
    UpdateServiceDto:
      type: object
      properties: {}
    CreateAppointmentDto:
      type: object
      properties: {}